requirements.txt
boto3
pika
python-dotenv
Flask
pyannote.audio
pydub